+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Token based login for node.js agent
+- Additional API functions allowing to query available agents, classes, methods
 
 ### Fixed